Update config_client to be more reusable
